% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarkPatientProfiles.R
\name{benchmarkPatientProfiles}
\alias{benchmarkPatientProfiles}
\title{Benchmark the PatientPrfiles packages for a certain cdm object}
\usage{
benchmarkPatientProfiles(cdm, prevalence = 0.8, seed = 123456)
}
\arguments{
\item{cdm}{A cdm_reference.}
}
\description{
Benchmark the PatientPrfiles packages for a certain cdm object
}
\examples{
\donttest{
library(PatientProfiles)

cdm <- mockPatientProfiles()

benchmarkPatientProfiles(cdm)
}

}
