% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDMBenchmark-package.R
\docType{package}
\name{CDMBenchmark-package}
\alias{CDMBenchmark}
\alias{CDMBenchmark-package}
\title{CDMBenchmark: Benchmark Packages Built Around the OMOP CDM}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
